+++ engine/src/main/java/dev/gamekit/ui/HBox.java
@@ -2,22 +2,9 @@ package dev.gamekit.ui;
 
 import dev.gamekit.utils.Reusable;
 
-import java.awt.*;
-
 public class HBox extends Container {
   @Override
-  protected void updateChildrenPositions() {
-    int xpos = padding.left;
-
-    for (var child : children) {
-      xpos += child.margin.left;
-      child.position.set(xpos, padding.top + child.margin.top);
-      xpos += child.size.width + child.margin.right;
-    }
-  }
-
-  @Override
-  protected Dimension computeChildrenSize() {
+  protected Size getContentSize() {
     int combinedWidth = 0;
     int tallestHeight = 0;
 
@@ -26,7 +13,18 @@ public class HBox extends Container {
       tallestHeight = Math.max(tallestHeight, child.size.height + child.margin.getVertical());
     }
 
-    Reusable.SIZE.setSize(combinedWidth, tallestHeight);
+    Reusable.SIZE.set(combinedWidth, tallestHeight);
     return Reusable.SIZE;
+  }
+
+  @Override
+  protected void updateContentPositions() {
+    int xpos = padding.left;
+
+    for (var child : children) {
+      xpos += child.margin.left;
+      child.position.set(xpos, padding.top + child.margin.top);
+      xpos += child.size.width + child.margin.right;
+    }
   }
 }